pyproject.toml
[project]
name = "azure-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-core>=1.39.0",
    "azure-storage-blob>=12.19.0",
    "azure-storage-queue>=12.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
